Split mod.rs into encode.rs, decode.rs and tests.rs
